Put ripple class in its .cpp file
